Created view model file for the app tab view and moved helper methods to said file.
